+++ apps/everest/pages/database-form/steps/second/second-step.const.ts
@@ -38,4 +38,4 @@ export const NODES_DB_TYPE_MAP: Record<DbType, string[]> = {
   [DbType.Mongo]: ['1', '3', '5'],
   [DbType.Mysql]: ['1', '3', '5'],
   [DbType.Postresql]: ['1', '2', '3'],
-}
+};
