+++ apps/everest/pages/database-form/useDatabaseFormDefaultValues.ts
@@ -19,7 +19,6 @@ import { useDbCluster } from '../../hooks/api/db-cluster/useDbCluster';
 import {
   // Backup,
   DbCluster,
-  ProxyExposeType,
 } from '../../types/dbCluster.types';
 import { dbEngineToDbType } from '../../utils/db';
 import { DB_WIZARD_DEFAULTS } from './database-form.constants';
@@ -28,103 +27,9 @@ import {
   DbWizardMode,
   DbWizardType,
 } from './database-form.types';
-import {
-  matchFieldsValueToResourceSize,
-  removeMeasurementValue,
-} from './steps/second/second-step.utils';
 import { useMonitoringInstancesList } from '../../hooks/api/monitoring/useMonitoringInstancesList';
-import {
-  MonitoringInstance,
-  MonitoringInstanceList,
-} from '../../types/monitoring.types';
-// import { getFormValuesFromCronExpression } from '../../components/time-selection/time-selection.utils';
-
-// EVEREST-334
-// const getBackupInfo = (backup: Backup) => {
-//   if (backup?.enabled) {
-//     const schedules = backup?.schedules;
-//     const firstSchedule = schedules && schedules[0];
-//     if (firstSchedule.schedule) {
-//       return {
-//         ...getFormValuesFromCronExpression(firstSchedule.schedule),
-//         [DbWizardFormFields.storageLocation]:
-//           { name: firstSchedule.backupStorageName } || null,
-//       };
-//     }
-//   }
-//   return {
-//     ...TIME_SELECTION_DEFAULTS,
-//     [DbWizardFormFields.storageLocation]:
-//       DB_WIZARD_DEFAULTS[DbWizardFormFields.storageLocation],
-//   };
-// };
-
-const getMonitoringInstanceValue = (
-  monitoringInstances: MonitoringInstanceList,
-  instanceName: string
-): MonitoringInstance | '' => {
-  if (monitoringInstances?.length) {
-    const monitoringInstance = monitoringInstances.find(
-      (item) => item.name === instanceName
-    );
-    if (monitoringInstance) return monitoringInstance;
-    return '';
-  }
-  return '';
-};
-
-export const DbClusterPayloadToFormValues = (
-  dbCluster: DbCluster,
-  monitoringInstances: MonitoringInstanceList,
-  mode: DbWizardMode
-): DbWizardType => {
-  // const backupInfo = getBackupInfo(dbCluster?.spec?.backup); // EVEREST-334
-
-  return {
-    // [DbWizardFormFields.backupsEnabled]: dbCluster?.spec?.backup?.enabled, // EVEREST-334
-    // [DbWizardFormFields.pitrEnabled]: true,
-    // [DbWizardFormFields.pitrTime]: '60',
-    // ...backupInfo, // EVEREST-334
-    [DbWizardFormFields.dbType]: dbEngineToDbType(
-      dbCluster?.spec?.engine?.type
-    ),
-    [DbWizardFormFields.dbName]:
-      mode === 'restoreFromBackup'
-        ? `restored-${dbCluster?.metadata?.name}`
-        : dbCluster?.metadata?.name,
-    [DbWizardFormFields.dbVersion]: dbCluster?.spec?.engine?.version || '',
-    [DbWizardFormFields.externalAccess]:
-      dbCluster?.spec?.proxy?.expose?.type === ProxyExposeType.external,
-    // [DbWizardFormFields.internetFacing]: true,
-    [DbWizardFormFields.engineParametersEnabled]:
-      !!dbCluster?.spec?.engine?.config,
-    [DbWizardFormFields.engineParameters]: dbCluster?.spec?.engine?.config,
-    [DbWizardFormFields.sourceRanges]: dbCluster?.spec?.proxy?.expose
-      ?.ipSourceRanges
-      ? dbCluster?.spec?.proxy?.expose?.ipSourceRanges.map((item) => ({
-          sourceRange: item,
-        }))
-      : [],
-    [DbWizardFormFields.monitoring]:
-      !!dbCluster?.spec?.monitoring?.monitoringConfigName,
-    [DbWizardFormFields.monitoringInstance]: getMonitoringInstanceValue(
-      monitoringInstances,
-      dbCluster?.spec?.monitoring?.monitoringConfigName
-    ),
-    [DbWizardFormFields.numberOfNodes]: `${dbCluster?.spec?.proxy?.replicas}`,
-    [DbWizardFormFields.resourceSizePerNode]:
-      matchFieldsValueToResourceSize(dbCluster),
-    [DbWizardFormFields.cpu]: +(dbCluster?.spec?.engine?.resources?.cpu || 0),
-    [DbWizardFormFields.disk]: removeMeasurementValue(
-      dbCluster?.spec?.engine?.storage?.size.toString()
-    ),
-    [DbWizardFormFields.memory]: removeMeasurementValue(
-      (dbCluster?.spec?.engine?.resources?.memory || 0).toString()
-    ),
-    [DbWizardFormFields.storageClass]:
-      dbCluster?.spec?.engine?.storage?.class || null,
-  };
-};
+import { useDbEngines } from '../../hooks/api/db-engines/useDbEngines';
+import { DbClusterPayloadToFormValues } from './database-form.utils';
 
 export const useDatabasePageDefaultValues = (
   mode: DbWizardMode
@@ -134,6 +39,7 @@ export const useDatabasePageDefaultValues = (
   dbClusterStatus: 'error' | 'idle' | 'loading' | 'success';
 } => {
   const { state } = useLocation();
+  const { data: dbEngines = [] } = useDbEngines();
   const { data, status } = useDbCluster(
     state?.selectedDbCluster,
     (mode === 'edit' || mode === 'restoreFromBackup') &&
@@ -159,6 +65,15 @@ export const useDatabasePageDefaultValues = (
         );
     } else setDefaultValues(DB_WIZARD_DEFAULTS);
   }, [data, monitoringInstances]);
+
+  useEffect(() => {
+    if (mode === 'new' && dbEngines.length > 0) {
+      const defaultDbType = dbEngineToDbType(dbEngines[0].type);
+      if (defaultDbType) {
+        defaultValues[DbWizardFormFields.dbType] = defaultDbType;
+      }
+    }
+  }, [dbEngines]);
 
   return { defaultValues, dbClusterData: data, dbClusterStatus: status };
 };
